requirements.txt

huggingface-hub==0.13.3
numpy==1.24.2
pandas==1.5.3
Pillow==7.1.2
scikit-learn==1.2.2
scipy==1.10.1
tensorboard==2.12.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.10.0
torchmetrics==0.11.4
tqdm==4.64.0
transformers==4.27.1
